rag-pipeline/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

mistralai==1.0.1

pymupdf==1.24.3

numpy==1.26.4

rank-bm25==0.2.2

pydantic==2.7.1
pydantic-settings==2.2.1

python-dotenv==1.0.1

slowapi==0.1.9

streamlit==1.35.0
requests==2.32.2
